rename plugin cisco_telemetry_gnmi to gnmi
