+++ README.md
@@ -164,7 +164,6 @@ For documentation on the latest development code see the [documentation index][d
 * [ceph](./plugins/inputs/ceph)
 * [cgroup](./plugins/inputs/cgroup)
 * [chrony](./plugins/inputs/chrony)
-* [cisco_telemetry_gnmi](./plugins/inputs/cisco_telemetry_gnmi)
 * [cisco_telemetry_mdt](./plugins/inputs/cisco_telemetry_mdt)
 * [clickhouse](./plugins/inputs/clickhouse)
 * [cloud_pubsub](./plugins/inputs/cloud_pubsub) Google Cloud Pub/Sub
@@ -197,6 +196,7 @@ For documentation on the latest development code see the [documentation index][d
 * [fireboard](/plugins/inputs/fireboard)
 * [fluentd](./plugins/inputs/fluentd)
 * [github](./plugins/inputs/github)
+* [gnmi](./plugins/inputs/gnmi)
 * [graylog](./plugins/inputs/graylog)
 * [haproxy](./plugins/inputs/haproxy)
 * [hddtemp](./plugins/inputs/hddtemp)
