+++ plugins/inputs/all/all.go
@@ -17,7 +17,6 @@ import (
 	_ "github.com/influxdata/telegraf/plugins/inputs/ceph"
 	_ "github.com/influxdata/telegraf/plugins/inputs/cgroup"
 	_ "github.com/influxdata/telegraf/plugins/inputs/chrony"
-	_ "github.com/influxdata/telegraf/plugins/inputs/cisco_telemetry_gnmi"
 	_ "github.com/influxdata/telegraf/plugins/inputs/cisco_telemetry_mdt"
 	_ "github.com/influxdata/telegraf/plugins/inputs/clickhouse"
 	_ "github.com/influxdata/telegraf/plugins/inputs/cloud_pubsub"
@@ -51,6 +50,7 @@ import (
 	_ "github.com/influxdata/telegraf/plugins/inputs/fireboard"
 	_ "github.com/influxdata/telegraf/plugins/inputs/fluentd"
 	_ "github.com/influxdata/telegraf/plugins/inputs/github"
+	_ "github.com/influxdata/telegraf/plugins/inputs/gnmi"
 	_ "github.com/influxdata/telegraf/plugins/inputs/graylog"
 	_ "github.com/influxdata/telegraf/plugins/inputs/haproxy"
 	_ "github.com/influxdata/telegraf/plugins/inputs/hddtemp"
